requirements.txt
fastapi==0.89.1
horapy==0.0.1
minio==7.1.13
numpy==1.24.1
opencv_python==4.7.0.68
PyJWT==2.6.0
requests==2.28.2
starlette==0.23.1
towhee==0.9.0
towhee.models==0.9.0
uvicorn==0.20.0
